Profiles' values are now floats; Units converting according the shared preferences

+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/dto/output/InsulinProfileOutput.kt
@@ -4,7 +4,7 @@ class InsulinProfileOutput(
     val profileName: String,
     val startTime: String,
     val endTime: String,
-    val glucoseObjective: Int,
-    val insulinSensitivityFactor: Int,
-    val carbohydrateRatio: Int
+    val glucoseObjective: Float,
+    val insulinSensitivityFactor: Float,
+    val carbohydrateRatio: Float
 )
